Port main's SAC-model redesign onto the graph-Voronoi env
main (rewritten history, force-pushed) independently redesigned the RL
observation/action/reward scheme and the SAC training loop while this
branch was switching the Voronoi cut itself to road-graph distance.
Bringing main's design over, adapted to keep the graph-based cut:

rl_env_voronoi_mw.py (state dimension + normalization):
- evaluate(a, return_cells=False): default 2-tuple (J, stds); pass
  return_cells=True for (J, stds, cnt, valid) used to build RL
  observations. Degenerate case (no segment captured) now returns a
  penalty J=1.0 instead of 0.0 (0.0 previously looked like a *perfect*
  score). within is now RMS of per-cell std instead of a plain mean.
- New obs_dim property (3*K+1) and _make_obs(): observation is no
  longer the raw K-dim `a`, but a concatenation of normalized a
  (a/a_bound), per-cell std (x10), a starved-cell mask, and J (x10) —
  each block scaled to roughly O(1) for the policy network.
- step(action): action is still a per-site Δa (K,), but now clipped to
  (-1,1) and scaled by _act_scale = 0.3*a_bound (bounded per-step move
  instead of an unscaled add). reward is now improvement-based
  (J_prev - J) instead of the absolute -J. step()/reset() return the
  built observation and an info dict (stds, J, cnt, n_valid).
- rho_max default 50.0 -> 2.0 (main's note: "within" exploits large
  rho_max by starving cells; 3-5 recommended, train.py overrides to 5).

Callers updated to the new evaluate() arity (2-tuple / return_cells
4-tuple, no more always-3-tuple): rl_env_vor_show.py, run_mw_test.py
(diagnostics rebuilt from cnt/valid since the old info dict is gone),
Comp_SA.py, Comp_Rand.py. Comp_SA.py also gets main's OBJECTIVE
"global" -> "within".

neural_networks.py / sac_algorithm.py / replay_buffer.py / train.py
were never touched independently on this branch, so taken wholesale
from main: deterministic policy eval mode, numerically stable tanh
log-prob correction, SAC.train_step()/@tf.function fused update reusing
log_probs across actor/alpha, float32 replay buffer storage, and
train.py's full rewrite (seeding, longer/shorter episode shape, reward
scaling, periodic deterministic eval, debug instrumentation). Added the
outputs_2008/*.csv data files train.py now points at.

+++ SAC/softlearning-master/softlearning/replay_buffer.py
@@ -6,11 +6,14 @@ class ReplayBuffer:
         self.ptr = 0
         self.size = 0
         
-        self.states = np.zeros((max_size, state_dim))
-        self.actions = np.zeros((max_size, action_dim))
-        self.rewards = np.zeros((max_size, 1))
-        self.next_states = np.zeros((max_size, state_dim))
-        self.dones = np.zeros((max_size, 1))
+        # 신경망이 어차피 float32 로 계산하므로 처음부터 float32 로 저장한다.
+        # float64 로 두면 메모리가 2배(94MB -> 47MB)이고, sample() 마다
+        # astype 으로 복사본을 한 번 더 만들게 된다.
+        self.states = np.zeros((max_size, state_dim), dtype=np.float32)
+        self.actions = np.zeros((max_size, action_dim), dtype=np.float32)
+        self.rewards = np.zeros((max_size, 1), dtype=np.float32)
+        self.next_states = np.zeros((max_size, state_dim), dtype=np.float32)
+        self.dones = np.zeros((max_size, 1), dtype=np.float32)
     
     def add(self, state, action, reward, next_state, done):
         self.states[self.ptr] = state
@@ -24,10 +27,11 @@ class ReplayBuffer:
     
     def sample(self, batch_size):
         indices = np.random.randint(0, self.size, batch_size)
+        # 저장이 이미 float32 라 astype 복사가 필요 없다 (복사 2회 -> 1회).
         return (
-            self.states[indices].astype('float32'),
-            self.actions[indices].astype('float32'),
-            self.rewards[indices].astype('float32'),
-            self.next_states[indices].astype('float32'),
-            self.dones[indices].astype('float32')
+            self.states[indices],
+            self.actions[indices],
+            self.rewards[indices],
+            self.next_states[indices],
+            self.dones[indices]
         )
